flask_api/requirements.txt
flask
requests
opencv-python
numpy
torch
torchvision
pillow
